refactor(run-protocol): move getRUN bootstrap into econ-behaviors
 - prune bootstrapRunLoC.js
