+++ packages/run-protocol/src/econ-behaviors.js
@@ -6,6 +6,7 @@ import {
   makeGovernedInvitation,
   CONTRACT_ELECTORATE,
   makeGovernedNat,
+  makeGovernedRatio,
 } from '@agoric/governance';
 import {
   CENTRAL_ISSUER_NAME,
@@ -15,16 +16,19 @@ import {
 import '@agoric/governance/exported.js';
 import '@agoric/vats/exported.js';
 
-import { makeGovernedTerms } from './vaultFactory/params.js';
-import { bootstrapRunLoC } from './bootstrapRunLoC.js';
+import { BLD_ISSUER_ENTRY } from '@agoric/vats/src/issuers.js';
+import { makeStakeReporter } from '@agoric/vats/src/my-lien.js';
+import {
+  makeElectorateParams,
+  makeGovernedTerms,
+} from './vaultFactory/params.js';
 
 import '../exported.js';
 
 import { PROTOCOL_FEE_KEY, POOL_FEE_KEY } from './vpool-xyk-amm/params.js';
 
-import { makeStakeReporter } from '../../vats/src/my-lien.js';
-import { BLD_ISSUER_ENTRY } from '../../vats/src/issuers.js';
 import { Collect } from './collect.js';
+import { CreditTerms } from './getRUN.js';
 
 const { entries, keys } = Object;
 
@@ -413,13 +417,85 @@ export const configureVaultFactoryUI = async ({
 };
 harden(configureVaultFactoryUI);
 
+/**
+ * @typedef { import('@agoric/eventual-send').Unpromise<T> } Unpromise<T>
+ * @template T
+ */
+/**
+ * @param {ERef<ZoeService>} zoe
+ * @param {ERef<TimerService>} timer
+ * @param { FeeMintAccess } feeMintAccess
+ * @param {Record<string, Installation>} installations
+ * @param {Object} terms
+ * @param {Ratio} terms.collateralPrice
+ * @param {Ratio} terms.collateralizationRatio
+ * @param {Issuer} stakeIssuer
+ *
+ * @typedef {Unpromise<ReturnType<typeof import('./getRUN.js').start>>} StartLineOfCredit
+ */
+export const bootstrapRunLoC = async (
+  zoe,
+  timer,
+  feeMintAccess,
+  installations,
+  { collateralPrice, collateralizationRatio },
+  stakeIssuer,
+) => {
+  const {
+    creatorFacet: electorateCreatorFacet,
+    instance: electorateInstance,
+  } = await E(zoe).startInstance(installations.electorate);
+  const poserInvitationP = E(electorateCreatorFacet).getPoserInvitation();
+  const [
+    initialPoserInvitation,
+    electorateInvitationAmount,
+  ] = await Promise.all([
+    poserInvitationP,
+    E(E(zoe).getInvitationIssuer()).getAmountOf(poserInvitationP),
+  ]);
+
+  // t.log({ electorateCreatorFacet, electorateInstance });
+  const main = harden({
+    [CreditTerms.CollateralPrice]: makeGovernedRatio(collateralPrice),
+    [CreditTerms.CollateralizationRatio]: makeGovernedRatio(
+      collateralizationRatio,
+    ),
+    ...makeElectorateParams(electorateInvitationAmount),
+  });
+
+  /** @type {{ publicFacet: GovernorPublic, creatorFacet: GovernedContractFacetAccess}} */
+  const governorFacets = await E(zoe).startInstance(
+    installations.governor,
+    {},
+    {
+      timer,
+      electorateInstance,
+      governedContractInstallation: installations.getRUN,
+      governed: harden({
+        terms: { main },
+        issuerKeywordRecord: { Stake: stakeIssuer },
+        privateArgs: { feeMintAccess, initialPoserInvitation },
+      }),
+    },
+    harden({ electorateCreatorFacet }),
+  );
+
+  const governedInstance = await E(governorFacets.creatorFacet).getInstance();
+  /** @type {ERef<StartLineOfCredit['publicFacet']>} */
+  const publicFacet = E(zoe).getPublicFacet(governedInstance);
+  const creatorFacet = E(governorFacets.creatorFacet).getCreatorFacet();
+
+  return { instance: governedInstance, publicFacet, creatorFacet };
+};
+harden(bootstrapRunLoC);
+
 /** @param {BootstrapPowers} powers */
 export const startGetRun = async ({
   consume: {
     zoe,
     // ISSUE: is there some reason Zoe shouldn't await this???
     feeMintAccess: feeMintAccessP,
-    governanceBundles,
+    getRUNBundle,
     agoricNames,
     bridgeManager,
     client,
@@ -429,7 +505,7 @@ export const startGetRun = async ({
 }) => {
   const [stakeName] = BLD_ISSUER_ENTRY;
 
-  const bundles = await governanceBundles;
+  const bundle = await getRUNBundle;
   const [
     feeMintAccess,
     bldIssuer,
@@ -446,7 +522,7 @@ export const startGetRun = async ({
     // TODO: manage string constants that need to match
     E(agoricNames).lookup('installation', 'contractGovernor'),
     E(agoricNames).lookup('installation', 'committee'),
-    E(zoe).install(bundles.getRUN),
+    E(zoe).install(bundle),
   ]);
   const collateralPrice = makeRatio(65n, runBrand, 100n, bldBrand); // arbitrary price
   const collateralizationRatio = makeRatio(5n, runBrand, 1n); // arbitrary raio
