+++ packages/run-protocol/test/test-getRUN.js
@@ -11,7 +11,8 @@ import { AmountMath, AssetKind, makeIssuerKit } from '@agoric/ertp';
 import { makeRatio } from '@agoric/zoe/src/contractSupport/index.js';
 import buildManualTimer from '@agoric/zoe/tools/manualTimer.js';
 
-import { bootstrapRunLoC, Collect } from '../src/bootstrapRunLoC.js';
+import { bootstrapRunLoC } from '../src/econ-behaviors.js';
+import { Collect } from '../src/collect.js';
 import * as testCases from './runLoC-test-case-sheet.js';
 
 const contractRoots = {
